+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.7.0] - 2023-11-28
+
+### Added
+
+- Support for QingTingFM (`.qta`) files.
+- Added CTR mode for AES encryption/decryption.
+
 ### Changed
 
 - Refactored AES code base with templates.
@@ -161,4 +168,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.6.3]: https://github.com/parakeet-rs/libparakeet/compare/v0.6.2...v0.6.3
 [0.6.4]: https://github.com/parakeet-rs/libparakeet/compare/v0.6.3...v0.6.4
 [0.6.5]: https://github.com/parakeet-rs/libparakeet/compare/v0.6.4...v0.6.5
-[0.x.0]: https://github.com/parakeet-rs/libparakeet/compare/v0.6.0...v0.x.0
+[0.7.0]: https://github.com/parakeet-rs/libparakeet/compare/v0.6.5...v0.7.0
+[0.x.0]: https://github.com/parakeet-rs/libparakeet/compare/v0.7.0...v0.x.0